Make sure that the default theme is the first one listed in themes.yml
We want the default theme to be Aubergine since that is the default that comes
with Slack. However, we were actually relying on the first theme in alphabetical
sorted order as the default theme, which ended up changing the default as themes
got added.

This patch always makes it the first entry in the themes.yml file