fix(Themes): Re-order and renamed GitKraken themes
Moved a GitKraken theme that was out of alphabetical order to the correct spot.

Due to this out-of-order theme, I mistakenly added another theme with the same name.

Renamed the theme I added that had the same name to be v2 because each theme is valid for GitKraken, but different enough to keep both.